+++ sphinxcontrib/needs/needs.py
@@ -37,7 +37,7 @@ else:
 
     logging.basicConfig()  # Only need to do this once
 
-VERSION = '0.3.8'
+VERSION = '0.3.9'
 
 DEFAULT_TEMPLATE_COLLAPSE = """
 .. _{{id}}: